refactoring stage matmul
