+++ crates/cubecl-linalg/src/matmul/components/stage/matmul/mod.rs
@@ -1,0 +1,8 @@
+mod base;
+pub mod plane;
+pub(super) mod shared;
+pub mod unit;
+
+pub use plane as plane_matmul;
+pub use shared::StageVectorization;
+pub use unit as unit_matmul;
