+++ crates/cubecl-linalg/src/matmul/components/stage/mod.rs
@@ -1,16 +1,14 @@
-pub mod plane_matmul;
-pub mod unit_matmul;
+mod matmul;
 
 mod base;
 mod event_listener;
 mod layout;
 mod reader;
-pub(super) mod shared;
 mod stage_memory;
 
 pub use base::*;
 pub use event_listener::*;
 pub use layout::*;
+pub use matmul::*;
 pub use reader::*;
-pub use shared::StageVectorization;
 pub use stage_memory::StageMemory;
